chore: consolidate util files
